Map work packages via template magic
Goal: Have Worker-classes that are comptible to
RoundRobinReinitScheduler also be usable by RoundRobinScheduler.

For this, we want to effectively implement lenses to extract information
from different work packages. In particular, we need to ignore all
session-related information for the plain RoundRobinScheduler.

In order to achieve this, we need to restructure our work-methods and
infer each trait with its own proper template meta-function.

Due to the verbose nature of templated code, this is slightly verbose.

Change-Id: I304bfac3ca9c3426a109d554d7656bfc2b7e4313