Add method to visit worker while set up
Because right now, `get_unique_identifer` requires an active connection
in hxcomm, we need a way to ensure the worker we visit is set up while
doing so.

Hence, we add visit_set_up_worker_const(...) methods to the schedulers.
These are simple forwards to visit_set_up_const on WorkerThread which in
turn ensure the worker is set up when the the visitor is applied.

Change-Id: I81ce36a3ed94b894f79399bb55a8bcba8b11ef09