Fix reinit not being executed for single user
If reinit is not ready yet for a single user, the session will be
switched but the reinit not performed. Once the requeued session is
executed, the reinit assumes the session is already established -> no
reinit is executed.

Hence, if we fail to perform the reinit, we need to reset the session
id.

Change-Id: If2dbf08ecd2d8bd4ead3eb9c9e07182d22e025b7