Indicate that reinit was performed
Currently, we use reinit in PyNN to perform the first setup phase while
the actual "realtime" part is performed via grenade.

That means that either both parts should run or none. We do this by
indicating that by enforcing the reinit. Here, the reinit will change
for every run, enforcing is fine.

However, in hxtorch we only need the reinit to be enforced once since it
is the calibration. Subsequent runs do not need to load the calibration
over and over again if the session is not switched.

Hence, after successful reinit we remove the "needed" tag (set via
reinit_enforce) so the reinit only gets applied once we relinquish
hardware control for a session.

Since ReinitStackEntry defaults to executing the supplied reinit
programs immediately, non-quiggeldy and quiggeldy execution modes remain
identical.

Change-Id: I30456317a0ee36467c2e5e73c2066f3dbf53b08f