Track OnDemandUploads with explicit ids
When using quiggeldy in stresstests, we found that there is a small
chance for glitches if a reint program takes too long to register prior
to executing a run. This is most evidend when only a single user is
performing PyNN-based experiments (as of 14.03.2021).

The robustness is improved by using three functions in the RCF interface
instead of two to implement on-request upload.

Previously:
* `bool notify()`:
  * Notifies the server of a new upload pending.
  * Blocks until server requests upload.
  * Returns bool:
    * true: client should upload
    * false: client should abort

* `void perform_upload(data)`:
  * Performs the actual upload.

Now: We generate a random `id` for every piece of data that we upload.
* `void notify(id)`:
  * Informs the server of a new reinit with given id.
  * Server should return "immediately".
  * After returning the client can assume the server knows about the
    reinit data.

* `bool pending(id)`:
  * Same functionality as `nofity()` in previous implementation, but
    called with id to ensure no other -- newer -- reinit program was
    registered.

* `perform_upload(data, id)`:
  * Same functionality as before, but called with explicit id.

Except for the change in the RCF interface (we add ids of type
`std::size_t`) there is no change in client code, but the server side
needs to be adapted.

The playground example (that serves as test) was adjusted.

NOTE: Because `RoundRobinReinintScheduler` is not yet adapted to this
change, building any dependents with only this change WILL FAIL!
However, in order to separate changes to the `OnDemandUpload` from the
changes to `RoundRobinReinintScheduler` this is acceptable for now.

Change-Id: Ie3aaf5943797736cd3d891b08d012a1ee8f76fd1