Fix reinit being applied all the time
Tracking reinit programs with an id makes the current submit_work api
superfluous.

Furthermore, there was a bug that once a reinit program was set to be
applied once it was applied all the time, which makes no difference for
pynn but slows down hxtorch tremendously. This change fixes that.

Change-Id: I9bb63c5b465fe218fae198acfc02bb2a63f7b35d