Attempt to fix quiggeldy stalls
Running a simple experiment in the loop, the quiggeldy connection is
showing an about 10 minute delay on occasion.

The reason appears to be that the main thread is waiting on a running
upload thread that refuses to join for about ten minutes.

The upload thread is designed to report it's progress about once per
second to a call back, which was not traced before.

Additionally, we explicitly request the current upload thread to stop
right before we attempt to join.

There is a fair chance this will not fix the issue, further datapoints
my @sschmitt are (probably) necessary.

Change-Id: I1a3b19f110dfdd24c97b127d4c1617965ee6c303